chore: Update _nav.blade.php with sticky top-0 class
This commit updates the _nav.blade.php file by adding the "sticky top-0" class to the main container div. This change ensures that the navigation bar remains fixed at the top of the page even when scrolling. It improves the user experience by providing easy access to the navigation menu at all times.